Refactor: dedupe relative time formatter
