+++ .agents/SESSIONS/2026-04-30.md
@@ -51,3 +51,60 @@ Started a new session after context reset. The session-start skill expected olde
 - `bun run typecheck` passed (`15 successful, 15 total`).
 - `bunx vitest run --reporter=dot` in `apps/desktop` passed (`62 files`, `352 tests`) and exited cleanly.
 - `bun run test` passed (`16 successful, 16 total`).
+
+## Session 3: Cleanup automation branch fix and relative time dedupe
+
+### System Flow
+
+```mermaid
+flowchart TD
+  A["Clean Up Code automation"] --> B["Codex worktree"]
+  B --> C{"Base branch correct?"}
+  C -->|wrong: master| D["Stop and rebase/recreate from develop"]
+  C -->|correct: develop| E["Find duplicate formatter code"]
+  E --> F["Shared formatRelativeTime helper"]
+  F --> G["Activity, Inbox, Overview, Issue Detail imports"]
+  G --> H["Commit and PR against develop"]
+```
+
+### Affected Components
+
+- `packages/shared`: new reusable relative-time formatter.
+- Desktop renderer: Activity, Inbox, Overview, and issue-detail relative time display.
+- Automation setup: clean-up automations now require `develop` except the Veref master-only exception.
+
+### What Was Done
+
+- Diagnosed the failed commit hook as a branch-base mismatch: the run started from `master`, but the shared hook expected a script present on `develop`.
+- Audited local clean-up automations and documented the Veref exception.
+- Reset the cleanup run onto current `origin/develop`.
+- Replaced four duplicate `timeAgo` implementations with one shared `formatRelativeTime` export.
+- Opened PR #92 against `develop`.
+
+### Key Decisions
+
+- Use `origin/develop` as the PR base so the cleanup PR does not carry unrelated local `develop` commits.
+- Keep the issue-detail `timeAgo` export as a compatibility alias to avoid touching the nested issue-detail tab imports.
+- Do not run full package typechecks in this worktree because `node_modules` is not installed; use Biome, hook validation, and a focused helper smoke check instead.
+
+### Files Changed
+
+- `packages/shared/src/format-relative-time.ts`
+- `packages/shared/src/index.ts`
+- `apps/desktop/src/renderer/components/ActivityView.tsx`
+- `apps/desktop/src/renderer/components/InboxView.tsx`
+- `apps/desktop/src/renderer/components/OverviewView.tsx`
+- `apps/desktop/src/renderer/components/issue-detail/helpers.ts`
+- `.agents/SESSIONS/2026-04-30.md`
+
+### Mistakes And Fixes
+
+- Mistake: the first cleanup commit was created from a master-based Codex worktree.
+- Fix: changed the automation source/default branch behavior, reset this branch onto `origin/develop`, and recreated the cleanup patch there.
+- Mistake: the first PR view temporarily included unrelated local develop commits while `origin/develop` was behind.
+- Fix: updated the branch after `origin/develop` advanced so the PR only carries this cleanup/session work.
+
+### Next Steps
+
+- Review PR #92.
+- Discard any stale master-based cleanup worktrees instead of continuing them.
