+++ packages/shared/src/format-relative-time.ts
@@ -1,0 +1,11 @@
+export function formatRelativeTime(input: string | number): string {
+  const timestamp = typeof input === 'number' ? input : new Date(input).getTime();
+  const diff = Math.max(0, Date.now() - timestamp);
+  const seconds = Math.floor(diff / 1000);
+  if (seconds < 60) return `${seconds}s ago`;
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes}m ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `${hours}h ago`;
+  return `${Math.floor(hours / 24)}d ago`;
+}
